backend/requirements/base.txt
Django>=5.0,<6.0
djangorestframework>=3.14
django-cors-headers>=4.0
psycopg2-binary>=2.9
redis>=4.5
django-redis>=5.4
celery>=5.3
pillow>=10.0
gunicorn>=21.0
whitenoise>=6.5
psutil>=5.9.0

# Media processing dependencies
moviepy>=1.0.3
pydub>=0.25.1
PyPDF2>=3.0.1